requirements.txt
six
protobuf==3.9.2
Flask
sentence-transformers
faiss-cpu
numpy
langchain
langchain-community
unstructured